# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GDT",
        "version": "1.0.0"
    },
    "paths": {
        "/api/clientesGetAll/{pagina}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Retorna lista de clientes",
                "operationId": "0a0f1b87afc8418faafedb5d58223e79",
                "parameters": [
                    {
                        "name": "página",
                        "in": "path",
                        "description": "Página que deseja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de clientes"
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Retorna um cliente pelo id",
                "operationId": "15931a6e5b60a2e66dfa2253a81dbc69",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do cliente que deseja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Exclui um cliente",
                "operationId": "51e0e6b7af9e46753298f685e5e523e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do cliente que deseja excluir",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/api/clienteTreinos/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Retorna o cliente e seus treinos",
                "operationId": "c8541e5cd65d07a666c2b14a5e110041",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id do cliente que deseja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente e seus treinos"
                    }
                }
            }
        },
        "/api/clientes/": {
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualiza um cliente",
                "operationId": "8f12bfc04151d16e9a85df2dff9272a3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "nome",
                                    "email",
                                    "telefone",
                                    "objetivo"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nome": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "35 99822 4443"
                                    },
                                    "objetivo": {
                                        "type": "string",
                                        "example": "Emagrecer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente atualizado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cadastra um cliente",
                "operationId": "b05062f6e14e36c64b19fd3fdb769550",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "telefone",
                                    "objetivo"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    },
                                    "telefone": {
                                        "type": "string",
                                        "example": "35 99822 4443"
                                    },
                                    "objetivo": {
                                        "type": "string",
                                        "example": "Emagrecer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente criado"
                    }
                }
            }
        },
...
    },
...
}